Name the main module "Main" as required
